requirements.txt
requests==2.27.1
lxml==4.7.1
selenium==4.1.0
PyJWT==2.3.0
